version bumb, move package and new package
Bitcoin, Binance, Android-studio, Discord,  version bump.
Atom, move package
Pot-desktop, new package
