+++ 0oldpackage/atom/actions.py
@@ -1,0 +1,15 @@
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+
+from pisi.actionsapi import get, pisitools, shelltools
+
+NoStrip = ["/usr"]
+IgnoreAutodep = True
+
+def setup():
+    shelltools.system("pwd")
+    shelltools.system("ar xf atom-amd64.deb")
+    shelltools.system("tar xf data.tar.xz")
+
+def install():
+    pisitools.insinto("/", "usr")
