+++ 0oldpackage/atom/pspec.xml
@@ -1,0 +1,95 @@
+<?xml version="1.0" ?>
+<!DOCTYPE PISI SYSTEM "https://pisilinux.org/projeler/pisi/pisi-spec.dtd">
+<PISI>
+    <Source>
+        <Name>atom</Name>
+        <Homepage>https://atom.io/</Homepage>
+        <Packager>
+            <Name>Pisi Linux Community</Name>
+            <Email>admins@pisilinux.org</Email>
+        </Packager>
+        <License>as-is</License>
+        <IsA>app:gui</IsA>
+        <PartOf>applications.development</PartOf>
+        <Summary>A hackable text editor.</Summary>
+        <Description>Atom is a text editor that's modern, approachable, yet hackable to the core—a tool you can customize to do anything but also use productively without ever touching a config file.</Description>
+        <Archive sha1sum="5c79f3615a9cab68b232618a5ed753c9efbe5426" type="binary">https://github.com/atom/atom/releases/download/v1.60.0/atom-amd64.deb</Archive>
+    </Source>
+    <Package>
+        <Name>atom</Name>
+        <RuntimeDependencies>
+            <Dependency>git</Dependency>
+            <Dependency>libgcrypt</Dependency>
+            <Dependency>gtk4</Dependency>
+            <Dependency>libnotify</Dependency>
+            <Dependency>nss</Dependency>
+            <Dependency>glib2</Dependency>
+            <Dependency>kde-cli-tools</Dependency>
+            <Dependency>xdg-utils</Dependency>
+            <Dependency>alsa-lib</Dependency>
+            <Dependency>mesa</Dependency>
+            <Dependency>libX11</Dependency>
+            <Dependency>libxcb</Dependency>
+            <Dependency>libXScrnSaver</Dependency>
+            <Dependency>libXtst</Dependency>
+            <Dependency>libxkbfile</Dependency>
+            <Dependency>curl</Dependency>
+        </RuntimeDependencies>
+        <Files>
+            <Path fileType="all">/usr</Path>         
+        </Files>
+     
+    </Package>
+
+    <History>
+        <Update release="7">
+            <Date>2024-01-16</Date>
+            <Version>1.60.0</Version>
+            <Comment>Version bump.</Comment>
+            <Name>Erkan IŞIK</Name>
+            <Email>erkanisik@pisilinux.org</Email>
+        </Update>
+          <Update release="6">
+            <Date>2021-07-16</Date>
+            <Version>1.57.0</Version>
+            <Comment>Version bump.</Comment>
+            <Name>Erkan IŞIK</Name>
+            <Email>erkanisik@pisilinux.org</Email>
+        </Update>
+         <Update release="5">
+            <Date>2021-04-01</Date>
+            <Version>1.55.0</Version>
+            <Comment>Version bump.</Comment>
+            <Name>Erkan IŞIK</Name>
+            <Email>erkanisik@pisilinux.org</Email>
+        </Update>
+        <Update release="4">
+            <Date>2019-07-12</Date>
+            <Version>1.38.2</Version>
+            <Comment>Version bump.</Comment>
+            <Name>Erkan IŞIK</Name>
+            <Email>erkanisik@pisilinux.org</Email>
+        </Update>
+        <Update release="3">
+            <Date>2018-02-19</Date>
+            <Version>1.24.0</Version>
+            <Comment>Version bump.</Comment>
+            <Name>Stefan Gronewold (groni)</Name>
+            <Email>groni@pisilinux.org</Email>
+        </Update>
+        <Update release="2">
+            <Date>2017-02-20</Date>
+            <Version>1.14.3</Version>
+            <Comment>Version bump.</Comment>
+            <Name>Stefan Gronewold (groni)</Name>
+            <Email>groni@pisilinux.org</Email>
+        </Update>
+        <Update release="1">
+            <Date>2016-05-6</Date>
+            <Version>1.8.0_beta3</Version>
+            <Comment>First release.</Comment>
+            <Name>Stefan Gronewold (groni)</Name>
+            <Email>groni@pisilinux.org</Email>
+        </Update>
+    </History>
+</PISI>
