--- development/atom/README.md
@@ -1,5 +1,0 @@
-For the Atom Package you need as Dependency:
-
-gconf and google-chrome
-
-you found gconf package at the main Repo you can  and here in pisilife-2 google-chrome package
