--- development/atom/files/atom.desktop
@@ -1,10 +1,0 @@
-[Desktop Entry]
-Name=Atom
-Comment=A hackable text editor for the 21st Century.
-GenericName=Text Editor
-Exec=/opt/atom/atom %U
-Icon=atom
-Type=Application
-StartupNotify=true
-Categories=TextEditor;Development;
-MimeType=text/plain;
